+++ src/pseudopeople/configuration/entities.py
@@ -4,6 +4,5 @@ class Keys:
     ROW_NOISE = "row_noise"  # second layer, eg <form>: row_noise: {...}
     COLUMN_NOISE = "column_noise"  # second layer, eg <form>: column_noise: {...}
     PROBABILITY = "probability"
-    ROW_NOISE_LEVEL = "row_noise_level"
     TOKEN_NOISE_LEVEL = "token_noise_level"
     AGE_MISWRITING_PERTURBATIONS = "possible_perturbations"